requirements.txt
elasticsearch~=5.0.0
elasticsearch-curator~=5.5.0
certifi
pyyaml
